Some refactoring of 103 to start with the problem.
