+++ src/main/java/eulerproject/level5/problem103/README.md
@@ -1,5 +1,5 @@
 
-### Problem 88
+### Problem 103
 
 #### Description
 Let S(A) represent the sum of elements in set A of size n. We shall call it a special sum set if for any two non-empty disjoint subsets, B and C, the following properties are true:
