+++ src/main/java/eulerproject/level5/problem105/readme.md
@@ -1,0 +1,35 @@
+# Problem 105
+
+## Description
+Let S(A) represent the sum of elements in set A of size n. We shall call it a special sum set if for any two non-empty 
+disjoint subsets, B and C, the following properties are true:
+
+* S(B) ≠ S(C); that is, sums of subsets cannot be equal.
+* If B contains more elements than C then S(B) > S(C).
+
+
+For example, {81, 88, 75, 42, 87, 84, 86, 65} is not a special sum set because 65 + 87 + 88 = 75 + 81 + 84, whereas 
+{157, 150, 164, 119, 79, 159, 161, 139, 158} satisfies both rules for all possible subset pair combinations and S(A) = 1286.
+
+Using sets.txt, a 4K text file with one-hundred sets containing seven to 
+twelve elements (the two examples given above are the first two sets in the file), identify all the special sum sets, 
+A1, A2, ..., Ak, and find the value of S(A1) + S(A2) + ... + S(Ak).
+
+NOTE: This problem is related to Problem 103 and Problem 106.
+
+## Solution
+
+## Ideas
+
+## Algorithm
+
+## Links
+[Problem 105](https://projecteuler.net/problem=105)
+
+
+[Problem 103](https://projecteuler.net/problem=103)
+[Problem 106](https://projecteuler.net/problem=106)
+Created by lgrzesik 10.05.20
+
+
+
