+++ src/main/java/eulerproject/pyeuler/timing.py
@@ -1,0 +1,9 @@
+import time
+
+
+def start():
+    return time.time_ns()
+
+
+def stop(start):
+    return (time.time_ns() - start) / (1e9)
